Struct trust_dns::authority::Authority

pub struct Authority {
    // some fields omitted
}

Authority is the storage method for all resource records

Methods

impl Authority

fn new(origin: Name, records: HashMap<Name, HashSet<Record>>, zone_type: ZoneType, allow_update: bool) -> Authority

fn get_origin(&self) -> &Name

fn get_zone_type(&self) -> ZoneType

fn get_soa(&self) -> Option<Record>

fn get_ns(&self) -> Option<Vec<Record>>

fn update(&mut self, update: &UpdateMessage) -> UpdateResult<()>

fn upsert(&mut self, name: Name, record: Record)

upserts into the resource vector Guarantees that SOA, CNAME only has one record, will implicitly update if they already exist

fn lookup(&self, name: &Name, rtype: RecordType, class: DNSClass) -> Option<Vec<Record>>