Module trust_dns::client

DNS Client associated classes for performing queries and other operations.

Structs

BasicClientHandle
Root ClientHandle implementaton returned by ClientFuture

ClientFuture
A DNS Client implemented over futures-rs.

ClientResponse
A future result of a Client Request

MemoizeClientHandle
A ClienHandle for memoized (cached) responses to queries.

RetryClientHandle
Can be used to reattempt a queries if they fail

SecureClientHandle
Performs DNSSec validation of all DNS responses from the wrapped DnsHandle

SecureSyncClient
A DNS client which will validate DNSSec records upon receipt

StreamHandle
The StreamHandle is the general interface for communicating with the DnsMultiplexer

SyncClient
The Client is abstracted over either trust_dns::tcp::TcpClientConnection or trust_dns::udp::UdpClientConnection.

Traits

Client
Client trait which implements basic DNS Client operations.

ClientConnection
Trait for client connections

ClientHandle
A trait for implementing high level functions of DNS.

ClientStreamHandle
Implementations of Sinks for sending DNS messages