Struct trust_dns::client::Client

pub struct Client<C: ClientConnection> {
    // some fields omitted
}

The Client is abstracted over either trust_dns::tcp::TcpClientConnection or trust_dns::udp::UdpClientConnection, usage of TCP or UDP is up to the user. Some DNS servers disallow TCP in some cases, so if TCP double check if UDP works.

Methods

impl<C: ClientConnection> Client<C>

fn new(client_connection: C) -> Client<C>

Creates a new DNS client with the specified connection type

Arguments

client_connection - the client_connection to use for all communication

fn with_trust_anchor(client_connection: C, trust_anchor: TrustAnchor) -> Client<C>

This variant allows for the trust_anchor to be replaced

Arguments

client_connection - the client_connection to use for all communication

trust_anchor - the set of trusted DNSKEY public_keys, by default this only contains the root public_key.

fn secure_query(&self, query_name: &Name, query_class: DNSClass, query_type: RecordType) -> ClientResult<Message>

DNSSec validating query, this will return an error if the requested records can not be validated against the trust_anchor.

When the resolver receives an answer via the normal DNS lookup process, it then checks to make sure that the answer is correct. Then starts with verifying the DS and DNSKEY records at the DNS root. Then use the DS records for the top level domain found at the root, e.g. 'com', to verify the DNSKEY records in the 'com' zone. From there see if there is a DS record for the subdomain, e.g. 'example.com', in the 'com' zone, and if there is use the DS record to verify a DNSKEY record found in the 'example.com' zone. Finally, verify the RRSIG record found in the answer for the rrset, e.g. 'www.example.com'.

Note As of now, this will not recurse on PTR or CNAME record responses, that is up to the caller.

Arguments

query_name - the label to lookup

query_class - most likely this should always be DNSClass::IN

query_type - record type to lookup

fn query(&self, name: &Name, query_class: DNSClass, query_type: RecordType) -> ClientResult<Message>

A classic DNS query, i.e. does not perform and DNSSec operations

Note As of now, this will not recurse on PTR or CNAME record responses, that is up to the caller.

Arguments

name - the label to lookup

query_class - most likely this should always be DNSClass::IN

query_type - record type to lookup