Struct trust_dns::client::SyncClient

pub struct SyncClient<CC> { /* fields omitted */ }

The Client is abstracted over either trust_dns::tcp::TcpClientConnection or trust_dns::udp::UdpClientConnection.

Usage of TCP or UDP is up to the user. Some DNS servers disallow TCP in some cases, so if TCP double check if UDP works.

Methods

impl<CC> SyncClient<CC> where
CC: ClientConnection,

pub fn new(conn: CC) -> Self

Creates a new DNS client with the specified connection type

Arguments

conn - the [ClientConnection] to use for all communication

pub fn with_signer(conn: CC, signer: Signer) -> Self

Creates a new DNS client with the specified connection type and a SIG0 signer.

This is necessary for signed update requests to update trust-dns-server entries.

Arguments

conn - the [ClientConnection] to use for all communication

signer - signer to use, this needs an associated private key