Struct trust_dns::op::header::Header

pub struct Header {
    // some fields omitted
}

Metadata for the Message struct.

RFC 1035, DOMAIN NAMES - IMPLEMENTATION AND SPECIFICATION, November 1987

4.1.1. Header section format

The header contains the following fields

                                   1  1  1  1  1  1
     0  1  2  3  4  5  6  7  8  9  0  1  2  3  4  5
    +--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+
    |                      ID                       |
    +--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+
    |QR|   Opcode  |AA|TC|RD|RA|ZZ|AD|CD|   RCODE   |  /// AD and CD from RFC4035
    +--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+
    |                    QDCOUNT / ZCOUNT           |
    +--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+
    |                    ANCOUNT / PRCOUNT          |
    +--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+
    |                    NSCOUNT / UPCOUNT          |
    +--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+
    |                    ARCOUNT / ADCOUNT          |
    +--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+

where

Z               Reserved for future use.  Must be zero in all queries
                and responses.

Methods

impl Header

fn new() -> Self

A default Header, not very useful.

fn len() -> usize

fn id(&mut self, id: u16) -> &mut Self

fn message_type(&mut self, message_type: MessageType) -> &mut Self

fn op_code(&mut self, op_code: OpCode) -> &mut Self

fn authoritative(&mut self, authoritative: bool) -> &mut Self

fn truncated(&mut self, truncated: bool) -> &mut Self

fn recursion_desired(&mut self, recursion_desired: bool) -> &mut Self

fn recursion_available(&mut self, recursion_available: bool) -> &mut Self

fn authentic_data(&mut self, authentic_data: bool) -> &mut Self

fn checking_disabled(&mut self, checking_disabled: bool) -> &mut Self

fn response_code(&mut self, response_code: ResponseCode) -> &mut Self

fn query_count(&mut self, query_count: u16) -> &mut Self

fn answer_count(&mut self, answer_count: u16) -> &mut Self

fn name_server_count(&mut self, name_server_count: u16) -> &mut Self

fn additional_count(&mut self, additional_count: u16) -> &mut Self

fn get_id(&self) -> u16

ID              A 16 bit identifier assigned by the program that
                generates any kind of query.  This identifier is copied
                the corresponding reply and can be used by the requester
                to match up replies to outstanding queries.

fn get_message_type(&self) -> MessageType

QR              A one bit field that specifies whether this message is a
                query (0), or a response (1).

fn get_op_code(&self) -> OpCode

OPCODE          A four bit field that specifies kind of query in this
                message.  This value is set by the originator of a query
                and copied into the response.  The values are: <see super::op_code>

fn is_authoritative(&self) -> bool

AA              Authoritative Answer - this bit is valid in responses,
                and specifies that the responding name server is an
                authority for the domain name in question section.

                Note that the contents of the answer section may have
                multiple owner names because of aliases.  The AA bit
                corresponds to the name which matches the query name, or
                the first owner name in the answer section.

fn is_truncated(&self) -> bool

TC              TrunCation - specifies that this message was truncated
                due to length greater than that permitted on the
                transmission channel.

fn is_recursion_desired(&self) -> bool

RD              Recursion Desired - this bit may be set in a query and
                is copied into the response.  If RD is set, it directs
                the name server to pursue the query recursively.
                Recursive query support is optional.

fn is_recursion_available(&self) -> bool

RA              Recursion Available - this be is set or cleared in a
                response, and denotes whether recursive query support is
                available in the name server.

fn is_authentic_data(&self) -> bool

RFC 4035, DNSSEC Resource Records, March 2005

3.1.6.  The AD and CD Bits in an Authoritative Response

  The CD and AD bits are designed for use in communication between
  security-aware resolvers and security-aware recursive name servers.
  These bits are for the most part not relevant to query processing by
  security-aware authoritative name servers.

  A security-aware name server does not perform signature validation
  for authoritative data during query processing, even when the CD bit
  is clear.  A security-aware name server SHOULD clear the CD bit when
  composing an authoritative response.

  A security-aware name server MUST NOT set the AD bit in a response
  unless the name server considers all RRsets in the Answer and
  Authority sections of the response to be authentic.  A security-aware
  name server's local policy MAY consider data from an authoritative
  zone to be authentic without further validation.  However, the name
  server MUST NOT do so unless the name server obtained the
  authoritative zone via secure means (such as a secure zone transfer
  mechanism) and MUST NOT do so unless this behavior has been
  configured explicitly.

  A security-aware name server that supports recursion MUST follow the
  rules for the CD and AD bits given in Section 3.2 when generating a
  response that involves data obtained via recursion.

fn is_checking_disabled(&self) -> bool

see is_authentic_data()

fn get_response_code(&self) -> u8

RCODE           Response code - this 4 bit field is set as part of
                responses.  The values have the following
                interpretation: <see super::response_code>

fn get_query_count(&self) -> u16

QDCOUNT         an unsigned 16 bit integer specifying the number of
                entries in the question section.

Return value

If this is a query, this will return the number of queries in the query section of the

fn get_answer_count(&self) -> u16

ANCOUNT         an unsigned 16 bit integer specifying the number of
                resource records in the answer section.

Return value

For query responses this is the number of records in the answer section, should be 0 for requests, for updates this is the count of prerequisite records.

fn get_name_server_count(&self) -> u16

for queries this is the nameservers which are authorities for the SOA of the Record for updates this is the update record count text NSCOUNT an unsigned 16 bit integer specifying the number of name server resource records in the authority records section.

Return value

For query responses this is the number of authorities, or nameservers, in the name server section, for updates this is the number of update records being sent.

fn get_additional_count(&self) -> u16

ARCOUNT         an unsigned 16 bit integer specifying the number of
                resource records in the additional records section.

Return value

This is the additional record section count, this section may include EDNS options.

fn clone(&self, query_count: u16, answer_count: u16, name_server_count: u16, additional_count: u16) -> Self

This is a specialized clone which clones all the fields but the counts handy for setting the count fields before sending over the wire.