Trait trust_dns::op::message::UpdateMessage

pub trait UpdateMessage: Debug {
    fn add_zone(&mut self, query: Query);
    fn add_pre_requisite(&mut self, record: Record);
    fn add_all_pre_requisites(&mut self, vector: &[Record]);
    fn add_update(&mut self, record: Record);
    fn add_all_updates(&mut self, vector: &[Record]);
    fn add_additional(&mut self, record: Record);
    fn get_zones(&self) -> &[Query];
    fn get_pre_requisites(&self) -> &[Record];
    fn get_updates(&self) -> &[Record];
    fn get_additional(&self) -> &[Record];
    fn get_sig0(&self) -> &[Record];
    fn sign(&mut self, signer: &Signer, inception_time: u32);
}

to reduce errors in using the Message struct as an Update, this will do the call throughs to properly do that.

Required Methods

fn add_zone(&mut self, query: Query)

fn add_pre_requisite(&mut self, record: Record)

fn add_all_pre_requisites(&mut self, vector: &[Record])

fn add_update(&mut self, record: Record)

fn add_all_updates(&mut self, vector: &[Record])

fn add_additional(&mut self, record: Record)

fn get_zones(&self) -> &[Query]

fn get_pre_requisites(&self) -> &[Record]

fn get_updates(&self) -> &[Record]

fn get_additional(&self) -> &[Record]

fn get_sig0(&self) -> &[Record]

returns the sig0, i.e. signed record, for verifying the sending and package integrity RFC 2535 section 4 this should be the final record in the additional section of the Message

RFC 2535 DNS Security Extensions March 1999

A DNS request may be optionally signed by including one or more SIGs at the end of the query. Such SIGs are identified by having a "type covered" field of zero. They sign the preceding DNS request message including DNS header but not including the IP header or any request SIGs at the end and before the request RR counts have been adjusted for the inclusions of any request SIG(s).

fn sign(&mut self, signer: &Signer, inception_time: u32)