Enum trust_dns::op::op_code::OpCode

pub enum OpCode {
    Query,
    Status,
    Notify,
    Update,
}

Operation code for queries, updates, and responses

RFC 1035, DOMAIN NAMES - IMPLEMENTATION AND SPECIFICATION, November 1987

OPCODE          A four bit field that specifies kind of query in this
                message.  This value is set by the originator of a query
                and copied into the response.  The values are:

                0               a standard query (QUERY)

                1               an inverse query (IQUERY)

                2               a server status request (STATUS)

                3-15            reserved for future use

Variants

Query

Query request RFC 1035

Status

Status message RFC 1035

Notify

Notify of change RFC 1996

Update

Update message RFC 2136