Enum trust_dns::op::response_code::ResponseCode

pub enum ResponseCode {
    NoError,
    FormErr,
    ServFail,
    NXDomain,
    NotImp,
    Refused,
    YXDomain,
    YXRRSet,
    NXRRSet,
    NotAuth,
    NotZone,
    BADVERS,
    BADSIG,
    BADKEY,
    BADTIME,
    BADMODE,
    BADNAME,
    BADALG,
    BADTRUNC,
    BADCOOKIE,
}

The status code of the response to a query.

RFC 1035, DOMAIN NAMES - IMPLEMENTATION AND SPECIFICATION, November 1987

RCODE           Response code - this 4 bit field is set as part of
                responses.  The values have the following
                interpretation:

                0               No error condition

                1               Format error - The name server was
                                unable to interpret the query.

                2               Server failure - The name server was
                                unable to process this query due to a
                                problem with the name server.

                3               Name Error - Meaningful only for
                                responses from an authoritative name
                                server, this code signifies that the
                                domain name referenced in the query does
                                not exist.

                4               Not Implemented - The name server does
                                not support the requested kind of query.

                5               Refused - The name server refuses to
                                perform the specified operation for
                                policy reasons.  For example, a name
                                server may not wish to provide the
                                information to the particular requester,
                                or a name server may not wish to perform
                                a particular operation (e.g., zone

                                transfer) for particular data.

                6-15            Reserved for future use.
 ```

Variants

NoError

No Error RFC 1035

FormErr

Format Error RFC 1035

ServFail

Server Failure RFC 1035

NXDomain

Non-Esistent Domain RFC 1035

NotImp

Not Implemented RFC 1035

Refused

Query Refused RFC 1035

YXDomain

Name Exists when it should not RFC 2136

YXRRSet

RR Set Exists when it should not RFC 2136

NXRRSet

RR Set that should exist does not RFC 2136

NotAuth

Server Not Authoritative for zone RFC 2136 or Not Authorized RFC 2845

NotZone

Name not contained in zone RFC 2136

BADVERS

Bad OPT Version RFC 6891

BADSIG

TSIG Signature Failure RFC 2845

BADKEY

Key not recognized RFC 2845

BADTIME

Signature out of time window RFC 2845

BADMODE

Bad TKEY Mode RFC 2930

BADNAME

Duplicate key name RFC 2930

BADALG

Algorithm not supported RFC 2930

BADTRUNC

Bad Truncation RFC 4635

BADCOOKIE

Bad/missing server cookie draft-ietf-dnsop-cookies

Methods

impl ResponseCode

fn low(&self) -> u8

returns the lower 4 bits of the response code (for the original header portion of the code)

fn high(&self)

returns the high 12 bits for the edns portion of the response code

fn from(high: u8, low: u8) -> ResponseCode

fn to_str(&self) -> &'static str