Struct trust_dns::op::Edns

pub struct Edns {
    // some fields omitted
}

Edns implements the higher level concepts for working with Edns as it is used to create or be created from OPT record data.

Methods

impl Edns

fn new() -> Self

fn get_rcode_high(&self) -> u8

fn get_version(&self) -> u8

fn is_dnssec_ok(&self) -> bool

fn get_max_payload(&self) -> u16

fn get_option(&self, code: &EdnsCode) -> Option<&EdnsOption>

fn get_options(&self) -> &OPT

fn set_rcode_high(&mut self, rcode_high: u8)

fn set_version(&mut self, version: u8)

fn set_dnssec_ok(&mut self, dnssec_ok: bool)

fn set_max_payload(&mut self, max_payload: u16)

fn set_option(&mut self, option: EdnsOption)