Struct trust_dns::op::LowerQuery

pub struct LowerQuery { /* fields omitted */ }

Identical to [trust_dns::op::Query], except that the Name is guaranteed to be in lower case form

Methods

impl LowerQuery

pub fn query(query: Query) -> Self

Create a new query from name and type, class defaults to IN

pub fn name(&self) -> &LowerName

QNAME           a domain name represented as a sequence of labels, where
                each label consists of a length octet followed by that
                number of octets.  The domain name terminates with the
                zero length octet for the null label of the root.  Note
                that this field may be an odd number of octets; no
                padding is used.

pub fn original(&self) -> &Query

Returns the original with the Names case preserved

pub fn query_type(&self) -> RecordType

QTYPE           a two octet code which specifies the type of the query.
                The values for this field include all codes valid for a
                TYPE field, together with some more general codes which
                can match more than one type of RR.

pub fn query_class(&self) -> DNSClass

QCLASS          a two octet code that specifies the class of the query.
                For example, the QCLASS field is IN for the Internet.