Trait trust_dns::op::MessageFinalizer

pub trait MessageFinalizer: 'static + Send + Sync {
    fn finalize_message(
        &self, 
        message: &Message, 
        current_time: u32
    ) -> Result<Vec<Record>, ProtoError>;
}

A trait for performing final ammendments to a Message before it is sent.

An example of this is a SIG0 signer, which needs the final form of the message, but then needs to attach additional data to the body of the message.

Required methods

fn finalize_message(
&self,
message: &Message,
current_time: u32
) -> Result<Vec<Record>, ProtoError>

The message taken in should be processed and then return [Record]s which should be appended to the additional section of the message.

Arguments

message - messge to process

current_time - the current time as specified by the system, it's not recommended to read the current time as that makes testing complicated.

Return

A vector to append to the additionals section of the message, sorted in the order as they should appear in the message.

Loading content...