Module trust_dns::op::update_message

Update related operations for Messages

Traits

UpdateMessage
To reduce errors in using the Message struct as an Update, this will do the call throughs to properly do that.

Functions

append
Appends a record to an existing rrset, optionally require the rrset to exis (atomicity depends on the server)

compare_and_swap
Compares and if it matches, swaps it for the new value (atomicity depends on the server)

create
Sends a record to create on the server, this will fail if the record exists (atomicity depends on the server)

delete_all
Deletes all records at the specified name

delete_by_rdata
Deletes a record (by rdata) from an rrset, optionally require the rrset to exist.

delete_rrset
Deletes an entire rrset, optionally require the rrset to exist.