Module trust_dns::rr

Resource record related components, e.g. Name aka label, Record, RData, ...

Modules

dns_class
class of DNS operations, in general always IN for internet

dnssec
dns security extension related modules

domain
Domain name associated types, suchas Name and Label.

rdata
All record data structures and related serialization methods

record_data
record data enum variants

record_type
record type definitions

resource
resource record implementation

zone
Reserved Zone and related information

Structs

Label
Labels are always stored as ASCII, unicode characters must be encoded with punycode

LowerName
them should be through references. As a workaround the Strings are all Rc as well as the array

Name
Them should be through references. As a workaround the Strings are all Rc as well as the array

Record
Resource records are storage value in DNS, into which all key/value pair data is stored.

RecordSet
Set of resource records associated to a name and type

RrKey
Accessor key for RRSets in the Authority.

Enums

DNSClass
The DNS Record class

RData
Record data enum variants

RecordType
The type of the resource record.

Traits

IntoName
Conversion into a Name

IntoRecordSet	Deprecated
Types which implement this can be converted into a RecordSet