Module trust_dns::rr::dnssec

dns security extension related modules

Modules

tbs
hash functions for DNSSec operations

Structs

Digest
A calculated digest value.

DnsSecError
The error type for dnssec errors that get returned in the crate

PublicKeyBuf
An owned variant of PublicKey

Signer
Use for performing signing and validation of DNSSec based components.

SupportedAlgorithms
Used to specify the set of SupportedAlgorithms between a client and server

TBS
Data To Be Signed.

TrustAnchor
The root set of trust anchors for validating DNSSec, anything in this set will be trusted

Enums

Algorithm
DNSSec signing and validation algorithms.

DigestType
This is the digest format for the

DnsSecErrorKind
The error kind for dnssec errors that get returned in the crate

KeyFormat
The format of the binary key

KeyPair
A public and private key pair, the private portion is not required.

Nsec3HashAlgorithm

Private
A tag type indicating that a key has private components.

Public
A tag type indicating that a key only has public components.

PublicKeyEnum
Variants of all know public keys

Traits

HasPrivate
A trait indicating that a key has private components.

HasPublic
A trait indicating that a key has public components.

PublicKey
PublicKeys implement the ability to ideally be zero copy abstractions over public keys for verifying signed content.

Verifier
Types which are able to verify DNS based signatures

Type Definitions

DnsSecResult
An alias for dnssec results returned by functions of this crate