Function trust_dns::rr::dnssec::tbs::rrset_tbs

pub fn rrset_tbs(
    name: &Name, 
    dns_class: DNSClass, 
    num_labels: u8, 
    type_covered: RecordType, 
    algorithm: Algorithm, 
    original_ttl: u32, 
    sig_expiration: u32, 
    sig_inception: u32, 
    key_tag: u16, 
    signer_name: &Name, 
    records: &[Record]
) -> Result<TBS, ProtoError>

Returns the to-be-signed serialization of the given record set.

Arguments

name - RRset record name

dns_class - DNSClass, i.e. IN, of the records

num_labels - number of labels in the name, needed to deal with *.example.com

type_covered - RecordType of the RRSet being hashed

algorithm - The Algorithm type used for the hashing

original_ttl - Original TTL is the TTL as specified in the SOA zones RRSet associated record

sig_expiration - the epoch seconds of when this hashed signature will expire

key_inception - the epoch seconds of when this hashed signature will be valid

signer_name - label of the etity responsible for signing this hash

records - RRSet to hash

Returns

the binary hash of the specified RRSet and associated information