Module trust_dns::rr::rdata

All record data structures and related serialization methods

Modules

a
IPv4 address record data

aaaa
IPv6 address record data

caa
allows a DNS domain name holder to specify one or more Certification Authorities (CAs) authorized to issue certificates for that domain.

dnskey
public key record data for signing zone records

ds
pointer record from parent zone to child zone for dnskey proof

key
public key record data for signing zone records

mx
mail exchange, email, record

name
Record type for all cname like records.

naptr
Dynamic Delegation Discovery System

nsec
negative cache proof for non-existence

nsec3
hashed negative cache proof for non-existence

nsec3param
parameters used for the nsec3 hash method

null
null record type, generally not used except as an internal tool for representing null data

openpgpkey
OPENPGPKEY records for OpenPGP public keys

opt
option record for passing protocol options between the client and server

sig
signature record for signing queries, updates, and responses

soa
start of authority record defining ownership and defaults for the zone

srv
service records for identify port mapping for specific services on a host

sshfp
SSHFP records for SSH public key fingerprints

tlsa
TLSA records for storing TLS certificate validation information

txt
text records for storing arbitrary data

Enums

DNSSECRData
Record data enum variants for DNSSEC-specific records.

DNSSECRecordType
The type of the resource record, for DNSSEC-specific records.