Enum trust_dns::rr::rdata::opt::EdnsOption

pub enum EdnsOption {
    DAU(SupportedAlgorithms),
    DHU(SupportedAlgorithms),
    N3U(SupportedAlgorithms),
    Unknown(u16, Vec<u8>),
}

options used to pass information about capabilities between client and server

note: Not all EdnsOptions are supported at this time.

http://www.iana.org/assignments/dns-parameters/dns-parameters.xhtml#dns-parameters-13

Variants

DAU

RFC 6975, DNSSEC Algorithm Understood

DHU

RFC 6975, DS Hash Understood

N3U

RFC 6975, NSEC3 Hash Understood

Unknown

Unknown, used to deal with unknown or unsupported codes

Methods

impl EdnsOption

fn len(&self) -> u16