Struct trust_dns::rr::rdata::soa::SOA

pub struct SOA {
    // some fields omitted
}

RFC 1035, DOMAIN NAMES - IMPLEMENTATION AND SPECIFICATION, November 1987

3.3.13. SOA RDATA format

    +--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+
    /                     MNAME                     /
    /                                               /
    +--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+
    /                     RNAME                     /
    +--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+
    |                    SERIAL                     |
    |                                               |
    +--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+
    |                    REFRESH                    |
    |                                               |
    +--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+
    |                     RETRY                     |
    |                                               |
    +--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+
    |                    EXPIRE                     |
    |                                               |
    +--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+
    |                    MINIMUM                    |
    |                                               |
    +--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+--+

where:

SOA records cause no additional section processing.

All times are in units of seconds.

Most of these fields are pertinent only for name server maintenance
operations.  However, MINIMUM is used in all query operations that
retrieve RRs from a zone.  Whenever a RR is sent in a response to a
query, the TTL field is set to the maximum of the TTL field from the RR
and the MINIMUM field in the appropriate SOA.  Thus MINIMUM is a lower
bound on the TTL field for all RRs in a zone.  Note that this use of
MINIMUM should occur when the RRs are copied into the response and not
when the zone is loaded from a master file or via a zone transfer.  The
reason for this provison is to allow future dynamic update facilities to
change the SOA RR with known semantics.

Methods

impl SOA

fn new(mname: Name, rname: Name, serial: u32, refresh: i32, retry: i32, expire: i32, minimum: u32) -> Self

Creates a new SOA record data.

Arguments

mname - the name of the master, primary, authority for this zone.

rname - the name of the responsible party for this zone, e.g. an email address.

serial - the serial number of the zone, used for caching purposes.

refresh - the amount of time to wait before a zone is resynched.

retry - the minimum period to wait if there is a failure during refresh.

expire - the time until this master is no longer authoritative for the zone.

minimum - no zone records should have time-to-live values less than this minimum.

Return value

The newly created SOA record data.

fn increment_serial(&mut self)

Increments the serial number by one

fn get_mname(&self) -> &Name

MNAME           The <domain-name> of the name server that was the
                original or primary source of data for this zone.

Return value

The domain-name of the name server that was the original or primary source of data for this zone, i.e. the master name server.

fn get_rname(&self) -> &Name

RNAME           A <domain-name> which specifies the mailbox of the
                person responsible for this zone.

Return value

A domain-name which specifies the mailbox of the person responsible for this zone, i.e. the responsible name.

fn get_serial(&self) -> u32

SERIAL          The unsigned 32 bit version number of the original copy
                of the zone.  Zone transfers preserve this value.  This
                value wraps and should be compared using sequence space
                arithmetic.

Return value

The unsigned 32 bit version number of the original copy of the zone. Zone transfers preserve this value. This value wraps and should be compared using sequence space arithmetic.

fn get_refresh(&self) -> i32

REFRESH         A 32 bit time interval before the zone should be
                refreshed.

Return value

A 32 bit time interval before the zone should be refreshed, in seconds.

fn get_retry(&self) -> i32

RETRY           A 32 bit time interval that should elapse before a
                failed refresh should be retried.

Return value

A 32 bit time interval that should elapse before a failed refresh should be retried, in seconds.

fn get_expire(&self) -> i32

EXPIRE          A 32 bit time value that specifies the upper limit on
                the time interval that can elapse before the zone is no
                longer authoritative.

Return value

A 32 bit time value that specifies the upper limit on the time interval that can elapse before the zone is no longer authoritative, in seconds

fn get_minimum(&self) -> u32

MINIMUM         The unsigned 32 bit minimum TTL field that should be
                exported with any RR from this zone.

Return value

The unsigned 32 bit minimum TTL field that should be exported with any RR from this zone.