Enum trust_dns::rr::record_type::RecordType

pub enum RecordType {
    A,
    AAAA,
    ANAME,
    ANY,
    AXFR,
    CAA,
    CNAME,
    IXFR,
    MX,
    NAPTR,
    NS,
    NULL,
    OPENPGPKEY,
    OPT,
    PTR,
    SOA,
    SRV,
    SSHFP,
    TLSA,
    TXT,
    DNSSEC(DNSSECRecordType),
    Unknown(u16),
    ZERO,
}

The type of the resource record.

This specifies the type of data in the RData field of the Resource Record

Variants
A

RFC 1035[1] IPv4 Address record

AAAA

RFC 3596[2] IPv6 address record

ANAME

ANAME draft-ietf-dnsop-aname

ANY

RFC 1035[1] All cached records, aka ANY

AXFR

RFC 1035[1] Authoritative Zone Transfer

CAA

RFC 6844 Certification Authority Authorization

CNAME

RFC 1035[1] Canonical name record

IXFR

RFC 1996 Incremental Zone Transfer

MX

RFC 1035[1] Mail exchange record

NAPTR

RFC 3403 Naming Authority Pointer

NS

RFC 1035[1] Name server record

NULL

RFC 1035[1] Null server record, for testing

OPENPGPKEY

RFC 7929 OpenPGP public key

OPT

RFC 6891 Option

PTR

RFC 1035[1] Pointer record

SOA

RFC 1035[1] and RFC 2308[9] Start of [a zone of] authority record

SRV

RFC 2782 Service locator

SSHFP

RFC 4255 SSH Public Key Fingerprint

TLSA

RFC 6698 TLSA certificate association

TXT

RFC 1035[1] Text record

DNSSEC(DNSSECRecordType)

A DNSSEC- or SIG(0)- specific record type.

These types are in DNSSECRecordType to make them easy to disable when crypto functionality isn't needed.

Unknown(u16)

Unknown Record type, or unsupported

ZERO

This corresponds to a record type of 0, unspecified

Methods

impl RecordType

pub fn is_any(self) -> bool

Returns true if this is an ANY

pub fn is_cname(self) -> bool

Returns true if this is a CNAME

pub fn is_srv(self) -> bool

Returns true if this is an SRV

pub fn is_ip_addr(self) -> bool

Returns true if this is an A or an AAAA record