Struct trust_dns::rr::LowerName

pub struct LowerName(_);

them should be through references. As a workaround the Strings are all Rc as well as the array

Methods

impl LowerName

pub fn new(name: &Name) -> Self

Create a new domain::LowerName, i.e. label

pub fn is_root(&self) -> bool

Returns true if there are no labels, i.e. it's empty.

In DNS the root is represented by .

Examples

use trust_dns::rr::{LowerName, Name};

let root = LowerName::from(Name::root());
assert_eq!(&root.to_string(), ".");

pub fn is_fqdn(&self) -> bool

Returns true if the name is a fully qualified domain name.

If this is true, it has effects like only querying for this single name, as opposed to building up a search list in resolvers.

warning: this interface is unstable and may change in the future

Examples

use std::str::FromStr;
use trust_dns::rr::{LowerName, Name};

let name = LowerName::from(Name::from_str("www").unwrap());
assert!(!name.is_fqdn());

let name = LowerName::from(Name::from_str("www.example.com").unwrap());
assert!(!name.is_fqdn());

let name = LowerName::from(Name::from_str("www.example.com.").unwrap());
assert!(name.is_fqdn());

pub fn base_name(&self) -> LowerName

Trims off the first part of the name, to help with searching for the domain piece

Examples

use std::str::FromStr;
use trust_dns::rr::{LowerName, Name};

let example_com = LowerName::from(Name::from_str("example.com").unwrap());
assert_eq!(example_com.base_name(), LowerName::from(Name::from_str("com.").unwrap()));
assert_eq!(LowerName::from(Name::from_str("com.").unwrap().base_name()), LowerName::from(Name::root()));
assert_eq!(LowerName::from(Name::root().base_name()), LowerName::from(Name::root()));

pub fn zone_of(&self, name: &Self) -> bool

returns true if the name components of self are all present at the end of name

Example

use std::str::FromStr;
use trust_dns::rr::{LowerName, Name};

let name = LowerName::from(Name::from_str("www.example.com").unwrap());
let zone = LowerName::from(Name::from_str("example.com").unwrap());
let another = LowerName::from(Name::from_str("example.net").unwrap());
assert!(zone.zone_of(&name));
assert!(!another.zone_of(&name));

pub fn num_labels(&self) -> u8

Returns the number of labels in the name, discounting *.

Examples

use std::str::FromStr;
use trust_dns::rr::{LowerName, Name};

let root = LowerName::from(Name::root());
assert_eq!(root.num_labels(), 0);

let example_com = LowerName::from(Name::from_str("example.com").unwrap());
assert_eq!(example_com.num_labels(), 2);

let star_example_com = LowerName::from(Name::from_str("*.example.com").unwrap());
assert_eq!(star_example_com.num_labels(), 2);

pub fn len(&self) -> usize

returns the length in bytes of the labels. '.' counts as 1

This can be used as an estimate, when serializing labels, they will often be compressed and/or escaped causing the exact length to be different.

pub fn is_empty(&self) -> bool

Returns true if the name is empty

pub fn emit_as_canonical(
&self,
encoder: &mut BinEncoder,
canonical: bool
) -> ProtoResult<()>

Emits the canonical version of the name to the encoder.

In canonical form, there will be no pointers written to the encoder (i.e. no compression).

pub fn is_wildcard(&self) -> bool

Pass through for Name::is_wildcard

pub fn into_wildcard(self) -> Self

Replaces the first label with the wildcard character, "*"