Enum trust_dns::serialize::binary::EncodeMode

pub enum EncodeMode {
    Verify,
    Normal,
}

In the Verify mode there maybe some things which are encoded differently, e.g. SIG0 records should not be included in the additional count and not in the encoded data when in Verify

Variants

Verify

Normal