Struct trust_dns::tcp::TcpStream

#[must_use = "futures do nothing unless polled"]

pub struct TcpStream<S> { /* fields omitted */ }

A Stream used for sending data to and from a remote DNS endpoint (client or server).

Methods

impl<S> TcpStream<S>

pub fn peer_addr(&self) -> SocketAddr

Returns the address of the peer connection.

impl TcpStream<TcpStream>

pub fn new<E>(
name_server: SocketAddr
) -> (Box<dyn Future<Item = TcpStream<TcpStream>, Error = Error> + 'static + Send>, BufStreamHandle) where
E: FromProtoError,

Creates a new future of the eventually establish a IO stream connection or fail trying.

Defaults to a 5 second timeout

Arguments

name_server - the IP and Port of the DNS server to connect to

pub fn with_timeout(
name_server: SocketAddr,
timeout: Duration
) -> (Box<dyn Future<Item = TcpStream<TcpStream>, Error = Error> + 'static + Send>, BufStreamHandle)

Creates a new future of the eventually establish a IO stream connection or fail trying

Arguments

name_server - the IP and Port of the DNS server to connect to

timeout - connection timeout

impl<S> TcpStream<S> where
S: AsyncRead + AsyncWrite,

pub fn from_stream(
stream: S,
peer_addr: SocketAddr
) -> (TcpStream<S>, BufStreamHandle)

Initializes a TcpStream with an existing tokio_tcp::TcpStream.

This is intended for use with a TcpListener and Incoming.

Arguments

stream - the established IO stream for communication

peer_addr - sources address of the stream

pub fn from_stream_with_receiver(
stream: S,
peer_addr: SocketAddr,
receiver: UnboundedReceiver<SerialMessage>
) -> TcpStream<S>

Wrapps a stream where a sender and receiver have already been established