Crate trust_dns

Trust-DNS is intended to be a fully compliant domain name server and client library.

Goals

Only safe Rust

All errors handled

Simple to manage servers

High level abstraction for clients

Secure dynamic update

New features for securing public information

Modules

authority
Module for Catalog of Authority zones which are responsible for storing RRSet records.

client
Use Client along with trust_dns::udp::UdpClientConnection or trust_dns::tcp::TcpClientConnection.

config
Configuration module for the server binary, named.

error
All defined errors for Trust-DNS

logger

op
Operations to send with a Client or server, e.g. Query, Message, or UpdateMessage can be used to gether to either query or update resource records sets.

rr
Resource record related components, e.g. Name aka label, Record, RData, ...

serialize
Contains serialization libraries for binary and text, txt.

server
Server component for hosting a domain name servers operations.

tcp
TCP protocol related components for DNS.

udp
UDP protocol related components for DNS.

Functions

version