Struct trust_dns::client::Client

pub struct Client<C: ClientConnection> {
    // some fields omitted
}

The Client is abstracted over either trust_dns::tcp::TcpClientConnection or trust_dns::udp::UdpClientConnection, usage of TCP or UDP is up to the user. Some DNS servers disallow TCP in some cases, so if TCP double check if UDP works.

Methods

impl<C: ClientConnection> Client<C>

fn new(client_connection: C) -> Client<C>

name_server to connect to with default port 53

fn secure_query(&self, query_name: &Name, query_class: DNSClass, query_type: RecordType) -> ClientResult<Message>

When the resolver receives an answer via the normal DNS lookup process, it then checks to make sure that the answer is correct. Then starts with verifying the DS and DNSKEY records at the DNS root. Then use the DS records for the top level domain found at the root, e.g. 'com', to verify the DNSKEY records in the 'com' zone. From there see if there is a DS record for the subdomain, e.g. 'example.com', in the 'com' zone, and if there is use the DS record to verify a DNSKEY record found in the 'example.com' zone. Finally, verify the RRSIG record found in the answer for the rrset, e.g. 'www.example.com'.

fn query(&self, name: &Name, query_class: DNSClass, query_type: RecordType) -> ClientResult<Message>