Enum trust_dns::rr::dnssec::Algorithm

pub enum Algorithm {
    RSASHA1,
    RSASHA256,
    RSASHA1NSEC3SHA1,
    RSASHA512,
}

Variants

RSASHA1

DO NOT USE, SHA1 is a compromised hashing function, it is here for backward compatability

RSASHA256

RSASHA1NSEC3SHA1

DO NOT USE, SHA1 is a compromised hashing function, it is here for backward compatability

RSASHA512

Methods

impl Algorithm

fn sign(&self, private_key: &PKey, data: &[u8]) -> Vec<u8>

fn verify(&self, public_key: &PKey, data: &[u8], signature: &[u8]) -> bool

fn from_u8(value: u8) -> DecodeResult<Self>

http://www.iana.org/assignments/dns-sec-alg-numbers/dns-sec-alg-numbers.xhtml

fn hash_len(&self) -> usize

length in bytes that the hash portion of this function will produce

fn public_key_from_vec(&self, public_key: &[u8]) -> DecodeResult<PKey>

fn public_key_to_vec(&self, public_key: &PKey) -> Vec<u8>