Struct trust_dns::rr::dnssec::Signer

pub struct Signer {
    // some fields omitted
}

Methods

impl Signer

fn new(algorithm: Algorithm, pkey: PKey, signer_name: Name) -> Self

fn get_algorithm(&self) -> Algorithm

fn get_signer_name(&self) -> &Name

fn calculate_key_tag(&self) -> u16

fn sign_message(&self, message: &Message) -> Vec<u8>

4.1.8.1 Calculating Transaction and Request SIGs

A response message from a security aware server may optionally contain a special SIG at the end of the additional information section to authenticate the transaction.

This SIG has a "type covered" field of zero, which is not a valid RR type. It is calculated by using a "data" (see Section 4.1.8) of the entire preceding DNS reply message, including DNS header but not the IP header and before the reply RR counts have been adjusted for the inclusion of any transaction SIG, concatenated with the entire DNS query message that produced this response, including the query's DNS header and any request SIGs but not its IP header. That is

data = full response (less transaction SIG) | full query

Verification of the transaction SIG (which is signed by the server host key, not the zone key) by the requesting resolver shows that the query and response were not tampered with in transit, that the response corresponds to the intended query, and that the response comes from the queried server.

A DNS request may be optionally signed by including one or more SIGs at the end of the query. Such SIGs are identified by having a "type covered" field of zero. They sign the preceding DNS request message including DNS header but not including the IP header or any request SIGs at the end and before the request RR counts have been adjusted for the inclusions of any request SIG(s).

WARNING: Request SIGs are unnecessary for any currently defined request other than update [RFC 2136, 2137] and will cause some old DNS servers to give an error return or ignore a query. However, such SIGs may in the future be needed for other requests.

Except where needed to authenticate an update or similar privileged request, servers are not required to check request SIGs.

NOTE: In classic RFC style, this is unclear, it implies that each SIG record is not included in the Additional record count, but this makes it more difficult to process and calculate more than one SIG0 record. Annoyingly, it means that the Header is signed with different material (i.e. additional record count - #SIG0 records), so the exact header sent is NOT the header being verified.

fn hash_rrset(&self, rrsig: &Record, records: &[Record]) -> Vec<u8>

name is the the name of the records in the rrset

fn verify(&self, hash: &[u8], signature: &[u8]) -> bool

verifies the hash matches the signature with the current key.