Struct trust_dns::rr::domain::Name

pub struct Name {
    // some fields omitted
}

TODO: all Names should be stored in a global "intern" space, and then everything that uses them should be through references.

Methods

impl Name

fn new() -> Self

fn label(self, label: &'static str) -> Self

fn with_labels(labels: Vec<String>) -> Self

fn add_label(&mut self, label: String) -> &mut Self

fn append(&mut self, other: &Self) -> &mut Self

fn base_name(&self) -> Option<Name>

Trims off the first part of the name, to help with searching for the domain piece TODO: This makes a copy right now, probably can do something better...

fn parse(local: &str, origin: Option<&Self>) -> ParseResult<Self>