Module trust_dns::rr::rdata

All record data structures and related serialization methods

Reexports

pub use self::dnskey::DNSKEY;

pub use self::ds::DS;

pub use self::mx::MX;

pub use self::nsec::NSEC;

pub use self::nsec3::NSEC3;

pub use self::nsec3param::NSEC3PARAM;

pub use self::null::NULL;

pub use self::opt::OPT;

pub use self::sig::SIG;

pub use self::srv::SRV;

pub use self::soa::SOA;

pub use self::txt::TXT;

Modules

a
IPv4 address record data

aaaa
IPv6 address record data

dnskey
public key record data for signing zone records

ds
pointer record from parent zone to child zone for dnskey proof

mx
mail exchange, email, record

name
Record type for all cname like records.

nsec
negative cache proof for non-existence

nsec3
hashed negative cache proof for non-existence

nsec3param
parameters used for the nsec3 hash method

null
null record type, generally not used except as an internal tool for representing null data

opt
option record for passing protocol options between the client and server

sig
signature record for signing queries, updates, and responses

soa
start of authority record defining ownership and defaults for the zone

srv
service records for identify port mapping for specific services on a host

txt
text records for storing arbitrary data