Enum trust_dns::rr::record_data::RData

pub enum RData {
    A(Ipv4Addr),
    AAAA(Ipv6Addr),
    CNAME(Name),
    DNSKEY(DNSKEY),
    DS(DS),
    MX(MX),
    NULL(NULL),
    NS(Name),
    NSEC(NSEC),
    NSEC3(NSEC3),
    NSEC3PARAM(NSEC3PARAM),
    OPT(OPT),
    PTR(Name),
    SIG(SIG),
    SOA(SOA),
    SRV(SRV),
    TXT(TXT),
}

Record data enum variants

RFC 1035, DOMAIN NAMES - IMPLEMENTATION AND SPECIFICATION, November 1987

3.3. Standard RRs

The following RR definitions are expected to occur, at least
potentially, in all classes.  In particular, NS, SOA, CNAME, and PTR
will be used in all classes, and have the same format in all classes.
Because their RDATA format is known, all domain names in the RDATA
section of these RRs may be compressed.

<domain-name> is a domain name represented as a series of labels, and
terminated by a label with zero length.  <character-string> is a single
length octet followed by that number of characters.  <character-string>
is treated as binary information, and can be up to 256 characters in
length (including the length octet).

Variants

A

AAAA

CNAME

DNSKEY

DS

MX

NULL

NS

NSEC

NSEC3

NSEC3PARAM

OPT

PTR

SIG

SOA

SRV

TXT

Methods

impl RData

fn parse(record_type: RecordType, tokens: &Vec<Token>, origin: Option<&Name>) -> ParseResult<Self>

fn read(decoder: &mut BinDecoder, record_type: RecordType, rdata_length: u16) -> DecodeResult<Self>

fn emit(&self, encoder: &mut BinEncoder) -> EncodeResult