Struct trust_dns::rr::resource::Record

pub struct Record {
    // some fields omitted
}

Methods

impl Record

fn new() -> Record

Creates a not very useful empty record, use the setters to build a more useful object There are no optional elements in this object, defaults are an empty name, type A, class IN, ttl of 0 and the 0.0.0.0 ip address.

fn name(&mut self, name: Name) -> &mut Self

fn add_name(&mut self, label: String) -> &mut Self

fn rr_type(&mut self, rr_type: RecordType) -> &mut Self

fn dns_class(&mut self, dns_class: DNSClass) -> &mut Self

fn ttl(&mut self, ttl: i32) -> &mut Self

fn rdata(&mut self, rdata: RData) -> &mut Self

fn get_name(&self) -> &Name

fn get_rr_type(&self) -> RecordType

fn get_dns_class(&self) -> DNSClass

fn get_ttl(&self) -> i32

fn get_rdata(&self) -> &RData